feat: a barebone De-Bruijn-Index boilerplate removal library
In the style of uniplate.

It's, as of yet, untested. I also need to add in examples. Good
examples could be λ-calculus, λ-calculus normal forms with hereditary
substitutions (for something with two syntactical categories), system
F (for something with two kinds of variables).

I probably also need to split the namespace into submodules. Probably
for substitutions and that's all. For the rest, let's figure out
what's needed on the spot.